+++ src/Config/Config.php
@@ -22,22 +22,24 @@ class Config extends ActiveRecordConfig
     const TABLE_NAME = "srusrenr_config";
     const PLUGIN_CLASS_NAME = ilSrUserEnrolmentPlugin::class;
     const KEY_ROLES = "roles";
+    const KEY_SHOW_ASSISTANTS = "show_assistants";
+    const KEY_SHOW_ENROLMENT_WORKFLOW = "show_enrolment_workflow";
     const KEY_SHOW_EXCEL_IMPORT = "show_excel_import";
     const KEY_SHOW_EXCEL_IMPORT_CONFIG = "show_excel_import_config";
     const KEY_SHOW_RESET_PASSWORD = "show_reset_password";
     const KEY_SHOW_RULES_ENROLL = "show_rules_enroll";
-    const KEY_SHOW_ENROLMENT_WORKFLOW = "show_enrolment_workflow";
     /**
      * @var array
      */
     protected static $fields
         = [
             self::KEY_ROLES                                               => [self::TYPE_JSON, []],
+            self::KEY_SHOW_ASSISTANTS                                     => [self::TYPE_BOOLEAN, false],
+            self::KEY_SHOW_ENROLMENT_WORKFLOW                             => [self::TYPE_BOOLEAN, false],
             self::KEY_SHOW_EXCEL_IMPORT                                   => [self::TYPE_BOOLEAN, false],
             self::KEY_SHOW_EXCEL_IMPORT_CONFIG                            => [self::TYPE_BOOLEAN, true],
             self::KEY_SHOW_RESET_PASSWORD                                 => [self::TYPE_BOOLEAN, false],
             self::KEY_SHOW_RULES_ENROLL                                   => [self::TYPE_BOOLEAN, false],
-            self::KEY_SHOW_ENROLMENT_WORKFLOW                             => [self::TYPE_BOOLEAN, false],
             ExcelImportFormGUI::KEY_COUNT_SKIP_TOP_ROWS                   => [self::TYPE_INTEGER, 0],
             ExcelImportFormGUI::KEY_CREATE_NEW_USERS                      => [self::TYPE_BOOLEAN, false],
             ExcelImportFormGUI::KEY_FIELDS                                => [