+++ src/EnrolmentWorkflow/Assistant/Assistants.php
@@ -1,0 +1,145 @@
+<?php
+
+namespace srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Assistant;
+
+use ActiveRecord;
+use arConnector;
+use ilSrUserEnrolmentPlugin;
+use srag\DIC\SrUserEnrolment\DICTrait;
+use srag\Plugins\SrUserEnrolment\Utils\SrUserEnrolmentTrait;
+
+/**
+ * Class Assistants
+ *
+ * @package srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Assistant
+ *
+ * @author  studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
+ */
+class Assistants extends ActiveRecord
+{
+
+    use DICTrait;
+    use SrUserEnrolmentTrait;
+    const TABLE_NAME = "srusrenr_ass";
+    const PLUGIN_CLASS_NAME = ilSrUserEnrolmentPlugin::class;
+
+
+    /**
+     * @inheritDoc
+     */
+    public function getConnectorContainerName() : string
+    {
+        return self::TABLE_NAME;
+    }
+
+
+    /**
+     * @inheritDoc
+     *
+     * @deprecated
+     */
+    public static function returnDbTableName() : string
+    {
+        return self::TABLE_NAME;
+    }
+
+
+    /**
+     * @var int
+     *
+     * @con_has_field    true
+     * @con_fieldtype    integer
+     * @con_length       8
+     * @con_is_notnull   true
+     * @con_is_primary   true
+     */
+    protected $user_id;
+    /**
+     * @var array
+     *
+     * @con_has_field    true
+     * @con_fieldtype    text
+     * @con_is_notnull   true
+     */
+    protected $assistants = [];
+
+
+    /**
+     * Assistants constructor
+     *
+     * @param int              $primary_key_value
+     * @param arConnector|null $connector
+     */
+    public function __construct(/*int*/ $primary_key_value = 0, arConnector $connector = null)
+    {
+        parent::__construct($primary_key_value, $connector);
+    }
+
+
+    /**
+     * @inheritDoc
+     */
+    public function sleep(/*string*/ $field_name)
+    {
+        $field_value = $this->{$field_name};
+
+        switch ($field_name) {
+            case "assistants":
+                return json_encode($field_value);
+
+            default:
+                return null;
+        }
+    }
+
+
+    /**
+     * @inheritDoc
+     */
+    public function wakeUp(/*string*/ $field_name, $field_value)
+    {
+        switch ($field_name) {
+            case "assistants":
+                return json_decode($field_value, true);
+
+            default:
+                return null;
+        }
+    }
+
+
+    /**
+     * @return int
+     */
+    public function getUserId() : int
+    {
+        return $this->user_id;
+    }
+
+
+    /**
+     * @param int $user_id
+     */
+    public function setUserId(int $user_id)/* : void*/
+    {
+        $this->user_id = $user_id;
+    }
+
+
+    /**
+     * @return array
+     */
+    public function getAssistants() : array
+    {
+        return $this->assistants;
+    }
+
+
+    /**
+     * @param array $assistants
+     */
+    public function setAssistants(array $assistants)/* : void*/
+    {
+        $this->assistants = $assistants;
+    }
+}